+++ Source/Document.js
@@ -1,0 +1,13 @@
+
+    
+export const queryAll = ( selector ) =>
+    document.querySelectorAll(selector);
+    
+export const query = ( selector ) =>
+    document.querySelector(selector);
+    
+export const create = ( type ) =>
+    document.createElement(type);
+    
+export const byId = ( id ) =>
+    document.getElementById(id);
