+++ Source/index.html
@@ -19,7 +19,9 @@
                 }
             }
         </script>
+        <script type = module src = Extensions.js></script>
         <script type = module src = App.js defer></script>
+        <script type = module src = Theme.js defer></script>
     </head>
     <body>
         <div>
